requirements.txt
discord.py
yt-dlp
PyNaCl
python-dotenv
google-api-python-client 
youtube-search-python